Scramjet.m from Homework 1
Kind of sets the structure for homework 1.  For the engine analysis, we break this down into the different components and bring in the other routines we have been working on... GHV inlet, Combustor, Nozzle.  Then we combine the engine drag thrust and lift forces along with lift and drag forces from a vehicle model and combine that into a flight sim... right?