Create ScramjetEngine.m
It's not complete yet, but this is the shell I was thinking we were creating for the Engine model.  Need to replace Combustor with Cal's function, and the nozzle with Marks.  I'm thinking that AJ is making a routine to look at the heat transfer. Then this gets pulled into the flight sim and combined with lift & drag to determine flight path.

Things to do include - upgrading to Cal and Marks models; then figuring out how to handle thrust summation and drag.